+++ Insurify.Population.DataAccess.Implementation/Models/CountryPopulationJson.cs
@@ -9,7 +9,7 @@ namespace Insurify.Population.DataAccess.Implementation.Models
         public string CountryCode { get; set; }
         [JsonPropertyName("Country Name")]
         public string CountryName { get; set; }
-        public float Value { get; set; }
+        public long Value { get; set; }
         public int Year { get; set; }
     }
 }
